New, repeated permutation results were generated; re-created data files from them. Also, found a small but important bug in weighted-regression.R; it's now fixed, and the data it generates (age-experience-tab.csv) is also created anew.
The new data always follows the (men - women) convention. This means I removed lots of plumbing code that was compensating for using it the other way round. Fortunately though, all the results are essentially identical to what we had before - so both the code semantics and the number of permutations were sufficient.

Finally, I put the prototype for Figure 4 in the figures folder (fig_4_test.pdf).